Codechange: made the codebase more modulair, decoupling a lot of code (#14)
Namespaces are now entities on their own, which register themself
to the WikiPage. This means no more hardcoded lists of namespaces
that are supported all over the place.
Adding new namespaces is now as simple as creating a new module
for it, and the existing namespaces will pick up on the new one.

This reduces code-complexity a lot, as there are no more
cross-references all over the place, and finding the right file
for a certain feature is now also more clear.