Add: change the license to AGPLv3 (#68)
All earlier versions will remain available under GPLv2.

As TrueWiki can be used as SaaS offering, switching to AGPL is
strongly advised, as that would make mandatory that anyone using
TrueWiki to offer their SaaS, still has to contribute back any
changes to this repository.

The whole idea of this licensing is that it is fine if you want
to use this software (for free), but have at least the decency to
contribute back any features or bugfixes you make. AGPL enforces
this a bit more, especially when used in a SaaS offering.