Fix: on "save page", push the new commit directly to GitHub
Otherwise there are too many edge-cases that can happen.
Some examples:
- If a PR is merged on GitHub, while we have changes pending. This
  would require conflict resolvement, which you really want to avoid
  if possible.
- During a new deployment, all pending changes are lost.

There is still a small window where things can go wrong (where
a PR is merged, but the GitHub Actions has not yet notified us
of a change), but this window is really small, and the impact is
at most a single page.

Fixes #50.