Add: move all git operations into another process
This drasticly speeds up the interface from a user-perspective,
and this is a fancy way of working around the GIL. Ideally we would
use a Git library that is async, which would also resolve this
issue, but with GitPython we are basically 4+ seconds waiting for
"git push" to finish. As that locks the GIL, no other request can
be handled while this is going on.

By moving the git operations to their own process, the main process
can safely continue to serve requests while in the background
git and GitHub are being updated. This is relative safe, as the
application itself doesn't care about git and/or GitHub. Locking
around the use of the out-of-process-git should prevent two
processes trying to access git and/or GitHub.

It is possible that if many people edit pages rapidly, that the
queue of things to commit gets long, and that commits start to
overlap. For example:
 create page -> modify -> modify -> modify
All within 5 seconds, will most likely create only two commits,
one with the create, and the rest with all the modifications
in one. It is, however, rather unlikely humans can go through
the interface within 5 seconds to make sane modifications. So
this is considered an accepted edge-case.