Add: don't run the metadata out-of-process and sys.intern() strings
The first prevents the need for twice the memory during any change
to the metadata. Instead, we yield on every page to allow aiohttp
to serve requests in between analyzing pages. This does increase
latency, but not by such an amount that it is annoying. And, it is
a rare event (rebuilding metadata), so it is unlikely people will
really notice.

sys.intern() reduces memory by 25% after reload, and by 10% when
there was no cache.