Add: run git commands via a "spawn" process
By default, Python uses "fork". But this requires twice the
memory of the main process, which can be more than a "spawn". For
git, we don't need access to any variable of the main process, so
a "spawn" is memory-wise cheaper. It does, however, take a bit
longer for a process to start. But this is okay, it is not that
bad.