Fix: only try to load a cache file from disk if the file exists (#84)
This could happen if the first time a page was loaded, was done
by someone who was logged in. This does update LAST_TIME_RENDERED
for browser if-modified-since caching, but does not write the
file to disk. If a non-logged-in user visit the page after, the
page is not on disk.