Fix: validate that cached files on disk are fresh (#85)
If a logged-in user generated a file, LAST_TIME_RENDERED was set,
but the cache on disk is not written. Now if there was an older
version of the cache still on disk, on next request from a visitor
that is not logged-in, the old cache would be served.

Validate that the time set in LAST_TIME_RENDERED is in fact also
the time of the cache-file on disk. If not, the file on disk is
stale and a new page should be rendered.