Add: .truewiki.yml support and move static into data-folder
This means that TrueWiki is now a stand-alone application that
can run different wiki's, depending on the data folder. CSS,
Javascript, and basic configuration comes from the data-folder
now.

Currently this means that running TrueWiki with an empty
data-folder doesn't give the best experience, as there is absolutely
no CSS loaded. Future commits will introduce a basic CSS to make
sure TrueWiki works out-of-the-box, but to keep things a bit clear
that is not part of this commit.