Fix #127: page_get_language() expects a valid URL, even for error pages
When someone access for example "/File", this is not a valid URL.
As such, an error page is generated which calls page_get_language().
This fails, as it is not a valid URL.

Take care of this by doing the minimal validation before calling
the function, and otherwise returning None.