Fix: always show all of the navigation bar, even on mobile (#239)
Otherwise people didn't find where to login or edit pages on
desktops with lower resolution.

Now when the window gets too small, the login bar is put below
the rest of the navigation. This uses up slightly more vertical
space, but at least keeps the navigation available for all resolutions.