Fix: git-storage-backend could try to commit several commits at once
There was an attempt to prevent that, but sadly, an Event was
used instead of a Lock. This meant that when the first commit
returned, all other pending were going through after that, instead
of one by one.

Lock is fair, as in, the first to acquire is the first to be released.