Fix: quickly adding and removing a page crashed the server
Because commits are queued, and if remotes are slow, they can
stack up. This makes it possible to add a page and remove one,
before it is committed. In result, when the commit is actually
being executed, it tried to add (or remove) a file that no longer
exists.

Solve this by simply ignoring these events, are they will be
fixed by the next commit in the queue. Sadly, it does mean we
lose a bit of history, but .. yeah, not much we can do about
that.