Fix: editing a page quickly after editing told you it was edited by someone else (#349)
This happened because the commit was still pending, so the nonce
was based on the old hash. Especially if there are a few commits
pending to be processed, this could easily happen.

By switching to modified-time of the file, like the local storage
backend does, we solve this problem. And it is safe, as TrueWiki
cannot scale out anyway for nodes that allow editing.