fix(cookies): allow more characters in cookie names, as defined by more recent RFCs
Python's http.cookies library is very strict, and throws an
exception when a cookie-name violates older RFC definitions of
the name. This is reflected back as a 500. But the user has no
indication that he was the cause of it.

More recent RFCs seem to allow for more characters, which is also
reflected in the http.cookies library, but not used. So switch to
that more relaxed definition of valid cookie name.

Still, illegal names should not return 500 in aiohttp, but that
is a problem for another time.