fix(cache): uploads are now cached "private, must-revalidate"
And favicon.ico is cached like all other static files.

The reason for this change is because people find it annoying to
wait for an updated image to appear. By allowing private caching,
but require revalidation, this is avoided. Sadly it also means
proxies will no longer cache the entry.
As a bonus, we can now extend the cache time to an hour, so
browsing the wiki is more smooth during that hour.

Also, only cache uploads that have a Last-Modified header,
as otherwise revalidation is pointless.